@@ -1,301 +1,11 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
-    <AspNetCoreHostingModel>InProcess</AspNetCoreHostingModel>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
-    <Compile Remove="Services\Binance\**" />
-    <Content Remove="Services\Binance\**" />
-    <EmbeddedResource Remove="Services\Binance\**" />
-    <None Remove="Services\Binance\**" />
-    <Content Remove="wwwroot\js\app.js" />
-    <Content Remove="wwwroot\images\checked.svg" />
-    <Content Remove="wwwroot\images\star.svg" />
-    <Content Remove="wwwroot\css\mobileStylesheet.css" />
-    <Content Remove="Views\Balances\_BalancesFilterPanel.cshtml" />
-    <Content Remove="Views\Balances\_BalancesFilterPanel.html" />
-    <Content Remove="Views\Home\PermissionDenied.cshtml" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Include="Views\Charts\AsksOnBids.cshtml" />
-    <None Include="Views\Charts\Charts.cshtml" />
-    <None Include="Views\Charts\DeltaOnTradeHistory.cshtml" />
-    <None Include="Views\Charts\_AsksOnBidsFilterPanel.cshtml" />
-    <None Include="Views\Charts\_DeltaOnTradeHistoryFilterPanel.cshtml" />
-    <None Include="Views\Data\Index.cshtml" />
-    <None Include="Views\Data\ShowOrderBookAsks.cshtml" />
-    <None Include="Views\Data\ShowOrderBookBids.cshtml" />
-    <None Include="Views\Data\ShowTradeDelta.cshtml" />
-    <None Include="Views\Data\ShowTradeHistory.cshtml" />
-    <None Include="Views\Data\_OrderBookFilterPanel.cshtml" />
-    <None Include="Views\Data\_TradeDeltaFilterPanel.cshtml" />
-    <None Include="Views\Data\_TradeHistoryFilterPanel.cshtml" />
-    <None Include="Views\Home\Index.cshtml" />
-    <None Include="Views\Home\Privacy.cshtml" />
-    <None Include="Views\Shared\Error.cshtml" />
-    <None Include="Views\Shared\_CookieConsentPartial.cshtml" />
-    <None Include="Views\Shared\_Layout.cshtml" />
-    <None Include="Views\Shared\_ValidationScriptsPartial.cshtml" />
-    <None Include="wwwroot\images\social\facebook.svg" />
-    <None Include="wwwroot\images\social\quora.svg" />
-    <None Include="wwwroot\images\social\reddit.svg" />
-    <None Include="wwwroot\images\social\slack.svg" />
-    <None Include="wwwroot\images\social\telegram.svg" />
-    <None Include="wwwroot\images\social\twitter.svg" />
-    <None Include="wwwroot\images\svg\accuracy.svg" />
-    <None Include="wwwroot\images\svg\analytics.svg" />
-    <None Include="wwwroot\images\svg\automations.svg" />
-    <None Include="wwwroot\images\svg\checked.svg" />
-    <None Include="wwwroot\images\svg\efficiency.svg" />
-    <None Include="wwwroot\images\svg\enchantment.svg" />
-    <None Include="wwwroot\images\svg\innovation.svg" />
-    <None Include="wwwroot\images\svg\lock.svg" />
-    <None Include="wwwroot\images\svg\logo-darktext-horizontal.svg" />
-    <None Include="wwwroot\images\svg\logo-whitetext-horizontal.svg" />
-    <None Include="wwwroot\images\svg\solar.svg" />
-    <None Include="wwwroot\images\svg\star.svg" />
-    <None Include="wwwroot\images\svg\vsn_bg.svg" />
-    <None Include="wwwroot\images\svg\why\exchanges\algorithm.svg" />
-    <None Include="wwwroot\images\svg\why\exchanges\execution.svg" />
-    <None Include="wwwroot\images\svg\why\exchanges\jigsaw.svg" />
-    <None Include="wwwroot\images\svg\why\exchanges\reports.svg" />
-    <None Include="wwwroot\images\svg\why\exchanges\star.svg" />
-    <None Include="wwwroot\images\timer-block\facebook.svg" />
-    <None Include="wwwroot\images\timer-block\telegram.svg" />
-    <None Include="wwwroot\images\timer-block\twitter.svg" />
-    <None Include="wwwroot\images\timer-block\vsn.svg" />
-    <None Include="wwwroot\js\app.js" />
-    <None Include="wwwroot\images\checked.svg" />
-    <None Include="wwwroot\images\star.svg" />
-    <None Include="wwwroot\js\summernote\lang\summernote-ar-AR.js" />
-    <None Include="wwwroot\js\summernote\lang\summernote-bg-BG.js" />
-    <None Include="wwwroot\js\summernote\lang\summernote-ca-ES.js" />
-    <None Include="wwwroot\js\summernote\lang\summernote-cs-CZ.js" />
-    <None Include="wwwroot\js\summernote\lang\summernote-da-DK.js" />
-    <None Include="wwwroot\js\summernote\lang\summernote-de-DE.js" />
-    <None Include="wwwroot\js\summernote\lang\summernote-el-GR.js" />
-    <None Include="wwwroot\js\summernote\lang\summernote-es-ES.js" />
-    <None Include="wwwroot\js\summernote\lang\summernote-es-EU.js" />
-    <None Include="wwwroot\js\summernote\lang\summernote-fa-IR.js" />
-    <None Include="wwwroot\js\summernote\lang\summernote-fi-FI.js" />
-    <None Include="wwwroot\js\summernote\lang\summernote-fr-FR.js" />
-    <None Include="wwwroot\js\summernote\lang\summernote-gl-ES.js" />
-    <None Include="wwwroot\js\summernote\lang\summernote-he-IL.js" />
-    <None Include="wwwroot\js\summernote\lang\summernote-hr-HR.js" />
-    <None Include="wwwroot\js\summernote\lang\summernote-hu-HU.js" />
-    <None Include="wwwroot\js\summernote\lang\summernote-id-ID.js" />
-    <None Include="wwwroot\js\summernote\lang\summernote-it-IT.js" />
-    <None Include="wwwroot\js\summernote\lang\summernote-ja-JP.js" />
-    <None Include="wwwroot\js\summernote\lang\summernote-ko-KR.js" />
-    <None Include="wwwroot\js\summernote\lang\summernote-lt-LT.js" />
-    <None Include="wwwroot\js\summernote\lang\summernote-lt-LV.js" />
-    <None Include="wwwroot\js\summernote\lang\summernote-mn-MN.js" />
-    <None Include="wwwroot\js\summernote\lang\summernote-nb-NO.js" />
-    <None Include="wwwroot\js\summernote\lang\summernote-nl-NL.js" />
-    <None Include="wwwroot\js\summernote\lang\summernote-pl-PL.js" />
-    <None Include="wwwroot\js\summernote\lang\summernote-pt-BR.js" />
-    <None Include="wwwroot\js\summernote\lang\summernote-pt-PT.js" />
-    <None Include="wwwroot\js\summernote\lang\summernote-ro-RO.js" />
-    <None Include="wwwroot\js\summernote\lang\summernote-ru-RU.js" />
-    <None Include="wwwroot\js\summernote\lang\summernote-sk-SK.js" />
-    <None Include="wwwroot\js\summernote\lang\summernote-sl-SI.js" />
-    <None Include="wwwroot\js\summernote\lang\summernote-sr-RS-Latin.js" />
-    <None Include="wwwroot\js\summernote\lang\summernote-sr-RS.js" />
-    <None Include="wwwroot\js\summernote\lang\summernote-sv-SE.js" />
-    <None Include="wwwroot\js\summernote\lang\summernote-ta-IN.js" />
-    <None Include="wwwroot\js\summernote\lang\summernote-th-TH.js" />
-    <None Include="wwwroot\js\summernote\lang\summernote-tr-TR.js" />
-    <None Include="wwwroot\js\summernote\lang\summernote-uk-UA.js" />
-    <None Include="wwwroot\js\summernote\lang\summernote-vi-VN.js" />
-    <None Include="wwwroot\js\summernote\lang\summernote-zh-CN.js" />
-    <None Include="wwwroot\js\summernote\lang\summernote-zh-TW.js" />
-    <None Include="wwwroot\js\summernote\plugin\databasic\summernote-ext-databasic.js" />
-    <None Include="wwwroot\js\summernote\plugin\hello\summernote-ext-hello.js" />
-    <None Include="wwwroot\js\summernote\plugin\specialchars\summernote-ext-specialchars.js" />
-    <None Include="wwwroot\js\summernote\summernote.min.js" />
-  </ItemGroup>
-
-
-  <ItemGroup>
-    <PackageReference Include="KK.DotNet.BackgroundTasks.Scheduled" Version="1.1.0" />
-    <PackageReference Include="Microsoft.AspNetCore.App">
-      <PrivateAssets Condition="'%(PackageReference.Version)' == ''">all</PrivateAssets>
-      <Publish Condition="'%(PackageReference.Version)' == ''">true</Publish>
-    </PackageReference>
-    <PackageReference Include="Microsoft.AspNetCore.Razor.Design" Version="2.2.0" PrivateAssets="All" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="2.2.6">
-      <PrivateAssets>all</PrivateAssets>
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers</IncludeAssets>
-    </PackageReference>
-    <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="2.2.4" />
-    <PackageReference Include="MySql.Data" Version="8.0.17" />
-    <PackageReference Include="MySql.Data.EntityFramework" Version="8.0.17" />
-    <PackageReference Include="MySql.Data.EntityFrameworkCore" Version="8.0.17" />
-    <PackageReference Include="MySql.Data.EntityFrameworkCore.Design" Version="8.0.17" />
     <PackageReference Include="Chart.js" Version="2.9.3" />
-  </ItemGroup>
-
-
-  <ItemGroup>
-    <Content Update="wwwroot\css\bootstrap-grid.css">
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
-    <Content Update="wwwroot\css\bootstrap-grid.css.map">
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
-    <Content Update="wwwroot\css\bootstrap-grid.min.css">
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
-    <Content Update="wwwroot\css\bootstrap-grid.min.css.map">
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
-    <Content Update="wwwroot\css\bootstrap-reboot.css">
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
-    <Content Update="wwwroot\css\bootstrap-reboot.css.map">
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
-    <Content Update="wwwroot\css\bootstrap-reboot.min.css">
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
-    <Content Update="wwwroot\css\bootstrap-reboot.min.css.map">
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
-    <Content Update="wwwroot\css\bootstrap.css">
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
-    <Content Update="wwwroot\css\bootstrap.css.map">
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
-    <Content Update="wwwroot\css\bootstrap.min.css">
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
-    <Content Update="wwwroot\css\bootstrap.min.css.map">
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
-    <Content Update="wwwroot\js\jquery-3.3.1.min.js">
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
-    <Content Update="wwwroot\js\jquery-latest.min.js">
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
-    <Content Update="wwwroot\js\jquery.tablesorter.js">
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
-    <Content Update="wwwroot\js\sorttable.js">
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
-    <Content Update="wwwroot\lib\bootstrap\.bower.json">
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
-    <Content Update="wwwroot\lib\bootstrap\dist\css\bootstrap-theme.css">
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
-    <Content Update="wwwroot\lib\bootstrap\dist\css\bootstrap-theme.css.map">
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
-    <Content Update="wwwroot\lib\bootstrap\dist\css\bootstrap-theme.min.css">
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
-    <Content Update="wwwroot\lib\bootstrap\dist\css\bootstrap-theme.min.css.map">
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
-    <Content Update="wwwroot\lib\bootstrap\dist\css\bootstrap.css">
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
-    <Content Update="wwwroot\lib\bootstrap\dist\css\bootstrap.css.map">
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
-    <Content Update="wwwroot\lib\bootstrap\dist\css\bootstrap.min.css">
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
-    <Content Update="wwwroot\lib\bootstrap\dist\css\bootstrap.min.css.map">
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
-    <Content Update="wwwroot\lib\bootstrap\dist\fonts\glyphicons-halflings-regular.eot">
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
-    <Content Update="wwwroot\lib\bootstrap\dist\fonts\glyphicons-halflings-regular.svg">
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
-    <Content Update="wwwroot\lib\bootstrap\dist\fonts\glyphicons-halflings-regular.ttf">
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
-    <Content Update="wwwroot\lib\bootstrap\dist\fonts\glyphicons-halflings-regular.woff">
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
-    <Content Update="wwwroot\lib\bootstrap\dist\fonts\glyphicons-halflings-regular.woff2">
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
-    <Content Update="wwwroot\lib\bootstrap\dist\js\bootstrap.js">
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
-    <Content Update="wwwroot\lib\bootstrap\dist\js\bootstrap.min.js">
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
-    <Content Update="wwwroot\lib\bootstrap\dist\js\npm.js">
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
-    <Content Update="wwwroot\lib\bootstrap\LICENSE">
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
-    <Content Update="wwwroot\lib\jquery-validation-unobtrusive\.bower.json">
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
-    <Content Update="wwwroot\lib\jquery-validation-unobtrusive\jquery.validate.unobtrusive.js">
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
-    <Content Update="wwwroot\lib\jquery-validation-unobtrusive\jquery.validate.unobtrusive.min.js">
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
-    <Content Update="wwwroot\lib\jquery-validation-unobtrusive\LICENSE.txt">
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
-    <Content Update="wwwroot\lib\jquery-validation\.bower.json">
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
-    <Content Update="wwwroot\lib\jquery-validation\dist\additional-methods.js">
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
-    <Content Update="wwwroot\lib\jquery-validation\dist\additional-methods.min.js">
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
-    <Content Update="wwwroot\lib\jquery-validation\dist\jquery.validate.js">
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
-    <Content Update="wwwroot\lib\jquery-validation\dist\jquery.validate.min.js">
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
-    <Content Update="wwwroot\lib\jquery-validation\LICENSE.md">
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
-    <Content Update="wwwroot\lib\jquery\.bower.json">
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
-    <Content Update="wwwroot\lib\jquery\dist\jquery.js">
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
-    <Content Update="wwwroot\lib\jquery\dist\jquery.min.js">
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
-    <Content Update="wwwroot\lib\jquery\dist\jquery.min.map">
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
-    <Content Update="wwwroot\lib\jquery\LICENSE.txt">
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
-  </ItemGroup>
-
-  <ItemGroup>
-    <Folder Include="wwwroot\json\" />
-    <Folder Include="Services\Balances\" />
-    <Folder Include="Controllers\Balances\" />
-    <Folder Include="ViewModels\Balances\" />
-    <Folder Include="Services\DataVisioAPI\" />
-    <Folder Include="Helpers\" />
   </ItemGroup>
 
   <ItemGroup>
