@@ -5,7 +5,7 @@
 
     <div class="content" style="min-height: 400px">
         <div class="container table_block" style="margin-top: 100px;">
-            <table class="sortable table table-bordered table-font" ">
+            <table class="sortable table table-bordered table-font" >
                 <thead>
                     <tr>
                         <th>Id</th>
